Updated battery remainig time calculations

+++ sources/hardware/arm/bq27441.h
@@ -100,7 +100,7 @@ class BQ27441 : public BatteryFuelGauge {
     int     getLevel() override;
     int     getHealth() override;
     bool    getIsCharging() override;
-    int     remainingLife() override;
+    float   remainingLife() override;
 
     int      getTemperatureC();  // Result in 1 Celsius
     uint16_t getFlags();
@@ -139,12 +139,12 @@ class BQ27441 : public BatteryFuelGauge {
     int     m_i2cDeviceId;
     int     m_i2cFd;
 
-    int  m_level                = 100;
-    int  m_health               = 100;
-    int  m_voltage              = 5;
-    int  m_averagePower         = 300;
-    bool m_isCharging           = false;
-    bool m_wasLowBatteryWarning = false;
-    int  m_remainingLife        = 0;
-    void updateBatteryValues();
+    int   m_level                = 100;
+    int   m_health               = 100;
+    int   m_voltage              = 5;
+    int   m_averagePower         = 300;
+    bool  m_isCharging           = false;
+    bool  m_wasLowBatteryWarning = false;
+    float m_remainingLife        = 0;
+    void  updateBatteryValues();
 };
